python-applic/requirements.txt
fastapi==0.117.1
uvicorn==0.37.0

prefect==3.4.19
pandas==2.3.2
openpyxl==3.1.5
cohere==5.17.0
litellm==1.77.1
supabase==2.18.1
beautifulsoup4==4.13.5
gspread==6.0.2
google-auth==2.30.0
requests==2.32.5
python-dotenv==1.0.1
urlextract==1.9.0
Crawl4AI==0.7.4
dacite==1.9.2
pydantic==2.11.9
